+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * new macro `begin*` that run their arguments in parallel if they async
 * add typecheck to rest of the functions
 * new `real?` and `pluck` functions
+### Bug fixes
+* fix `find` and `filter` to check if function from argument return nil
 
 ## 0.11.2
 ### Breaking